refactor: 플레이어와 딜러를 추상화하는 Participant 추가 및 생성자가 없는 클래스를 object로 변경
