+++ src/main/kotlin/blackjack/Application.kt
@@ -1,10 +1,7 @@
 package blackjack
 
 import blackjack.controller.BlackJackController
-import blackjack.model.CardDeck
 
 fun main() {
-    BlackJackController(
-        CardDeck(),
-    ).startGame()
+    BlackJackController.startGame()
 }
