--- src/main/kotlin/blackjack/model/CardDrawer.kt
@@ -1,5 +1,0 @@
-package blackjack.model
-
-interface CardDrawer {
-    fun drawCard(generateCard: () -> Card): PickingState
-}
