+++ src/main/kotlin/blackjack/model/Participant.kt
@@ -1,0 +1,10 @@
+package blackjack.model
+
+interface Participant {
+    fun drawCard(generateCard: () -> Card): PickingState
+
+    fun drawUntilSatisfaction(
+        generateCard: () -> Card,
+        printCards: (GameInfo) -> Unit,
+    )
+}
